perf(core): Make `PlatformLocation` tree-shakable
Convert `PlatformLocation` into a tree-shakable provider.
