+++ integration/_payload-limits.json
@@ -3,7 +3,7 @@
     "master": {
       "uncompressed": {
         "runtime": 1497,
-        "main": 161490,
+        "main": 158490,
         "polyfills": 45399
       }
     }
@@ -21,7 +21,7 @@
     "master": {
       "uncompressed": {
         "runtime": 1440,
-        "main": 128448,
+        "main": 125448,
         "polyfills": 45340
       }
     }
