+++ integration/side-effects/snapshots/forms/esm5.js
@@ -2,7 +2,7 @@ import "tslib";
 
 import "@angular/core";
 
-import "@angular/platform-browser";
+import "@angular/common";
 
 import "rxjs";
