+++ integration/side-effects/snapshots/router/esm2015.js
@@ -5,5 +5,3 @@ import "@angular/core";
 import "rxjs";
 
 import "rxjs/operators";
-
-import "@angular/platform-browser";
