+++ integration/side-effects/snapshots/router/esm5.js
@@ -7,5 +7,3 @@ import "@angular/core";
 import "rxjs";
 
 import "rxjs/operators";
-
-import "@angular/platform-browser";
