+++ packages/common/src/location/location_strategy.ts
@@ -6,8 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {InjectionToken} from '@angular/core';
-import {LocationChangeListener} from './platform_location';
+import {Inject, Injectable, InjectionToken, Optional, ɵɵinject} from '@angular/core';
+import {DOCUMENT} from '../dom_tokens';
+import {LocationChangeListener, PlatformLocation} from './platform_location';
+import {joinWithSlash, normalizeQueryParams} from './util';
 
 /**
  * Enables the `Location` service to read route state from the browser's URL.
@@ -26,6 +28,7 @@ import {LocationChangeListener} from './platform_location';
  *
  * @publicApi
  */
+@Injectable({providedIn: 'root', useFactory: provideLocationStrategy})
 export abstract class LocationStrategy {
   abstract path(includeHash?: boolean): string;
   abstract prepareExternalUrl(internal: string): string;
@@ -35,6 +38,13 @@ export abstract class LocationStrategy {
   abstract back(): void;
   abstract onPopState(fn: LocationChangeListener): void;
   abstract getBaseHref(): string;
+}
+
+export function provideLocationStrategy(platformLocation: PlatformLocation) {
+  // See #23917
+  const location = ɵɵinject(DOCUMENT).location;
+  return new PathLocationStrategy(
+      ɵɵinject(PlatformLocation as any), location && location.origin || '');
 }
 
 
@@ -62,3 +72,82 @@ export abstract class LocationStrategy {
  * @publicApi
  */
 export const APP_BASE_HREF = new InjectionToken<string>('appBaseHref');
+
+/**
+ * @description
+ * A {@link LocationStrategy} used to configure the {@link Location} service to
+ * represent its state in the
+ * [path](https://en.wikipedia.org/wiki/Uniform_Resource_Locator#Syntax) of the
+ * browser's URL.
+ *
+ * If you're using `PathLocationStrategy`, you must provide a {@link APP_BASE_HREF}
+ * or add a base element to the document. This URL prefix that will be preserved
+ * when generating and recognizing URLs.
+ *
+ * For instance, if you provide an `APP_BASE_HREF` of `'/my/app'` and call
+ * `location.go('/foo')`, the browser's URL will become
+ * `example.com/my/app/foo`.
+ *
+ * Similarly, if you add `<base href='/my/app'/>` to the document and call
+ * `location.go('/foo')`, the browser's URL will become
+ * `example.com/my/app/foo`.
+ *
+ * @usageNotes
+ *
+ * ### Example
+ *
+ * {@example common/location/ts/path_location_component.ts region='LocationComponent'}
+ *
+ * @publicApi
+ */
+@Injectable()
+export class PathLocationStrategy extends LocationStrategy {
+  private _baseHref: string;
+
+  constructor(
+      private _platformLocation: PlatformLocation,
+      @Optional() @Inject(APP_BASE_HREF) href?: string) {
+    super();
+
+    if (href == null) {
+      href = this._platformLocation.getBaseHrefFromDOM();
+    }
+
+    if (href == null) {
+      throw new Error(
+          `No base href set. Please provide a value for the APP_BASE_HREF token or add a base element to the document.`);
+    }
+
+    this._baseHref = href;
+  }
+
+  onPopState(fn: LocationChangeListener): void {
+    this._platformLocation.onPopState(fn);
+    this._platformLocation.onHashChange(fn);
+  }
+
+  getBaseHref(): string { return this._baseHref; }
+
+  prepareExternalUrl(internal: string): string { return joinWithSlash(this._baseHref, internal); }
+
+  path(includeHash: boolean = false): string {
+    const pathname =
+        this._platformLocation.pathname + normalizeQueryParams(this._platformLocation.search);
+    const hash = this._platformLocation.hash;
+    return hash && includeHash ? `${pathname}${hash}` : pathname;
+  }
+
+  pushState(state: any, title: string, url: string, queryParams: string) {
+    const externalUrl = this.prepareExternalUrl(url + normalizeQueryParams(queryParams));
+    this._platformLocation.pushState(state, title, externalUrl);
+  }
+
+  replaceState(state: any, title: string, url: string, queryParams: string) {
+    const externalUrl = this.prepareExternalUrl(url + normalizeQueryParams(queryParams));
+    this._platformLocation.replaceState(state, title, externalUrl);
+  }
+
+  forward(): void { this._platformLocation.forward(); }
+
+  back(): void { this._platformLocation.back(); }
+}
