+++ packages/common/src/location/platform_location.ts
@@ -6,7 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {InjectionToken} from '@angular/core';
+import {Inject, Injectable, InjectionToken, ɵɵinject} from '@angular/core';
+import {getDOM} from '../dom_adapter';
+import {DOCUMENT} from '../dom_tokens';
+
 /**
  * This class should not be used directly by an application developer. Instead, use
  * {@link Location}.
@@ -29,6 +32,11 @@ import {InjectionToken} from '@angular/core';
  *
  * @publicApi
  */
+@Injectable({
+  providedIn: 'platform',
+  // See #23917
+  useFactory: useBrowserPlatformLocation
+})
 export abstract class PlatformLocation {
   abstract getBaseHrefFromDOM(): string;
   abstract getState(): unknown;
@@ -50,6 +58,10 @@ export abstract class PlatformLocation {
   abstract forward(): void;
 
   abstract back(): void;
+}
+
+export function useBrowserPlatformLocation() {
+  return ɵɵinject(BrowserPlatformLocation);
 }
 
 /**
@@ -75,3 +87,80 @@ export interface LocationChangeEvent {
  * @publicApi
  */
 export interface LocationChangeListener { (event: LocationChangeEvent): any; }
+
+
+
+/**
+ * `PlatformLocation` encapsulates all of the direct calls to platform APIs.
+ * This class should not be used directly by an application developer. Instead, use
+ * {@link Location}.
+ */
+@Injectable({
+  providedIn: 'platform',
+  // See #23917
+  useFactory: createBrowserPlatformLocation,
+})
+export class BrowserPlatformLocation extends PlatformLocation {
+  public readonly location !: Location;
+  private _history !: History;
+
+  constructor(@Inject(DOCUMENT) private _doc: any) {
+    super();
+    this._init();
+  }
+
+  // This is moved to its own method so that `MockPlatformLocationStrategy` can overwrite it
+  /** @internal */
+  _init() {
+    (this as{location: Location}).location = getDOM().getLocation();
+    this._history = getDOM().getHistory();
+  }
+
+  getBaseHrefFromDOM(): string { return getDOM().getBaseHref(this._doc) !; }
+
+  onPopState(fn: LocationChangeListener): void {
+    getDOM().getGlobalEventTarget(this._doc, 'window').addEventListener('popstate', fn, false);
+  }
+
+  onHashChange(fn: LocationChangeListener): void {
+    getDOM().getGlobalEventTarget(this._doc, 'window').addEventListener('hashchange', fn, false);
+  }
+
+  get href(): string { return this.location.href; }
+  get protocol(): string { return this.location.protocol; }
+  get hostname(): string { return this.location.hostname; }
+  get port(): string { return this.location.port; }
+  get pathname(): string { return this.location.pathname; }
+  get search(): string { return this.location.search; }
+  get hash(): string { return this.location.hash; }
+  set pathname(newPath: string) { this.location.pathname = newPath; }
+
+  pushState(state: any, title: string, url: string): void {
+    if (supportsState()) {
+      this._history.pushState(state, title, url);
+    } else {
+      this.location.hash = url;
+    }
+  }
+
+  replaceState(state: any, title: string, url: string): void {
+    if (supportsState()) {
+      this._history.replaceState(state, title, url);
+    } else {
+      this.location.hash = url;
+    }
+  }
+
+  forward(): void { this._history.forward(); }
+
+  back(): void { this._history.back(); }
+
+  getState(): unknown { return this._history.state; }
+}
+
+export function supportsState(): boolean {
+  return !!window.history.pushState;
+}
+export function createBrowserPlatformLocation() {
+  return new BrowserPlatformLocation(ɵɵinject(DOCUMENT));
+}
