+++ packages/core/src/di/util.ts
@@ -10,6 +10,7 @@ import {Type} from '../interface/type';
 import {ReflectionCapabilities} from '../reflection/reflection_capabilities';
 import {getClosureSafeProperty} from '../util/property';
 
+import {resolveForwardRef} from './forward_ref';
 import {injectArgs, ɵɵinject} from './injector_compatibility';
 import {ClassSansProvider, ConstructorSansProvider, ExistingSansProvider, FactorySansProvider, StaticClassSansProvider, ValueProvider, ValueSansProvider} from './interface/provider';
 
@@ -32,7 +33,7 @@ export function convertInjectableProviderToFactory(
     return () => valueProvider.useValue;
   } else if ((provider as ExistingSansProvider).useExisting) {
     const existingProvider = (provider as ExistingSansProvider);
-    return () => ɵɵinject(existingProvider.useExisting);
+    return () => ɵɵinject(resolveForwardRef(existingProvider.useExisting));
   } else if ((provider as FactorySansProvider).useFactory) {
     const factoryProvider = (provider as FactorySansProvider);
     return () => factoryProvider.useFactory(...injectArgs(factoryProvider.deps || EMPTY_ARRAY));
@@ -43,7 +44,7 @@ export function convertInjectableProviderToFactory(
       const reflectionCapabilities = new ReflectionCapabilities();
       deps = reflectionCapabilities.parameters(type);
     }
-    return () => new classProvider.useClass(...injectArgs(deps));
+    return () => new (resolveForwardRef(classProvider.useClass))(...injectArgs(deps));
   } else {
     let deps = (provider as ConstructorSansProvider).deps;
     if (!deps) {
