--- packages/platform-browser/src/browser/location/history.ts
@@ -1,11 +1,0 @@
-/**
- * @license
- * Copyright Google Inc. All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-export function supportsState(): boolean {
-  return !!window.history.pushState;
-}
